+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - DATA: Use "Burmese" script for language "Mon"
 - DATA: Use "Baybayin" script name instead of "Tagalog (Baybayin, Alibata)"
 - TWEAK: Make sure `Orthography.base_chars` and `Orthography.aux_chars` return no duplicates for decomposed character sequences
+- TWEAK: Define `Languages`, `Language` and `Orthography` as module top level exports for easier importing, e.g. now: `from hyperglot import Language`
 
 ## 0.6.3 (08.04.2024)
 - FIX: Set correct default values for `Language.status` and `Orthography.preferred_as_group` and provide validation and tests for these.